Fix video playback regression in Safari 12
Playback of files with audio was broken on Safari 12 (as tested with
Safari Technology preview version 59) in the default mwembed-based
player.

This is due to Safari for Mac adopting anti-autoplay technology earlier
deployed in iOS; we were working around it but only on iOS. Changed it
to unconditionally use the workaround (which is to start an AudioContext
in the playback handler, then reuse it in the actual playback).

Bug: T197914
Change-Id: Idcfd85adfe6f49cab332ae911b172cf457abecd7
(cherry picked from commit 5f9c5e6d470a5c8c43e3b4fe13964d862bd0554d)